requirements.txt
semantic-kernel
langchain
langchain-community
langchain-openai
python-dotenv